// src/app/api/model/cu-policy-setting-model.ts
/**
 * CrediFlow.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CUPolicySettingModel { 
    policyId?: string | null;
    storeIds?: Array<string> | null;
    effectiveFrom: string;
    effectiveTo?: string | null;
    earlySettlementPenaltyRate?: number;
    latePaymentPenaltyRate?: number;
    latePaymentStartDay?: number;
    badDebtStartDay?: number;
    insuranceDiscountRate?: number;
    warningDays?: Array<number> | null;
}